third_party/blink/web_tests/virtual/policy-container/external/wpt/referrer-policy/generic/inheritance/iframe-inheritance-history-about-srcdoc-expected.txt

This is a testharness.js-based test.
FAIL History navigation reuses original policy. assert_equals: History navigation reuses original policy. expected (undefined) undefined but got (string) "http://web-platform.test:8001/"
PASS New srcdoc iframe uses new policy.
Harness: the test ran to completion.
